Fixed a "side effect" bug with latest `beginning-of-defun' and `end-of-defun' advices.  They caused font-lock, which uses beginning/end of defun to force a reparse.  Thanks to "Eric M. Ludlam" <zappo@ultranet.com> for pointing this.
Improved consistency with standard behaviour of `beginning-of-defun'
(go to the beginning of the line where the defun starts) and
`end-of-defun' (go to the beginning of the line following the end of
the defun).

Added useful advices for `narrow-to-defun', `mark-defun' and
`c-mark-function'.

Advices are enabled when the functions are called interactively and
`senator-minor-mode' is enabled.